Split widget: timeline row
